Fix: review comments
